Simplify library structure (#1)
* simplify directory tree

The scoped folders here are pointless and just make the library tedious to navigate.

* simplify gem namespaces

Per the last commit, we don't need all of these namespaces. Removed the useless intermediate modules and updated all paths accordingly.

* remove duplicate pry requirement

When authoring a gem, all gems should be specified in your gemspec. Pry was also already in the gemspec, throwing a warning upon installation.

* prefer sandboxed installation in bin/setup

* restructure specs to match updated library structure